CIRISVoice/requirements.txt
wyoming==1.5.0
aiohttp==3.9.1
pydantic==2.5.0
pyyaml==6.0.1
openai==1.6.0
google-cloud-speech==2.22.0
google-cloud-texttospeech==2.15.0
google-auth==2.26.2
python-dotenv==1.0.0

# CIRIS SDK - installed separately from local source
# ciris-sdk>=0.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
